Feat/git workflow (#9)
* Add tag-triggered editor release workflow

* Refactor Linux runtime library CMake helper

* Fix Linux build errors

+++ engine/src/core/types.cpp
@@ -23,7 +23,7 @@ const char* mental::core::resultToString(mental::core::Result res)
     default:
     {
       static char buf[24];
-      snprintf(buf, sizeof(buf), "Unknown (%d)", res);
+      snprintf(buf, sizeof(buf), "Unknown (%d)", static_cast<int>(res));
       return buf;
     }
   }
